+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## master
 
+* Changed the default value of `start` to be `undefined`
+* Changed the default value of `step` to be `undefined`
+* Added missing `animate` attribute, default to `true`.
+
 ## 0.1.1 2015-03-09
 
 * Moved rimraf dependency to devDependencies